Optimize Bloom Filter Hashing loop strength and base hashing
Optimize Bloom Filter performance by:
1. Invoking gxhash::gxhash64 directly in get_base_hash to avoid GxHasher builder overhead and unnecessary i64 casts.
2. Implementing loop strength reduction in insert/contains to use incremental additive hashing instead of full multiplication on each loop iteration.

+++ src/hypertrie/src/bloom_filter.rs
@@ -49,7 +49,7 @@ impl BloomFilter {
 
     #[inline(always)]
     fn get_base_hash(&self, item: &[u8]) -> u64 {
-        gxhash::gxhash64(item, SEED as i64)
+        gxhash::gxhash64(item, SEED)
     }
 }
